requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==1.10.14

python-dotenv==1.0.1
requests==2.31.0

google-generativeai==0.7.2
pinecone-client==3.2.2

PyPDF2==3.0.1
numpy==1.26.4
tiktoken==0.6.0

python-multipart==0.0.6
